# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ledgers api endpoints for the Accounts Ledgers.\nIt serves as Accounts Ledgers.",
        "title": "Ledgers Clean Architecture.",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:8089",
    "basePath": "/api/v1",
    "paths": {
        "/blogs/": {
            "get": {
                "description": "Show all blogs registered",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blogs"
                ],
                "summary": "List all the Blogs",
                "operationId": "Fetch Blog",
                "parameters": [
                    {
                        "type": "string",
                        "description": "1",
                        "name": "Page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "10",
                        "name": "ItemsPerPage",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.listBlogsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/blogs/create-blog/": {
            "post": {
                "description": "Create a blog",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blogs"
                ],
                "summary": "Create a blog",
                "operationId": "Create a blog",
                "parameters": [
                    {
                        "description": "Create blog request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.createBlogRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/v1.createBlogResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/blogs/{id}": {
            "get": {
                "description": "Show a single blog registered",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blogs"
                ],
                "summary": "Fetch single blog by ID",
                "operationId": "Single blog",
                "parameters": [
                    {
                        "type": "string",
                        "description": "blog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.singleBlogResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "db.Blog": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "descriptions": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userRole": {
                    "type": "string"
                }
            }
        },
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "v1.createBlogRequestBody": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        },
        "v1.createBlogResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "v1.listBlogsResponse": {
            "type": "object",
            "properties": {
                "blogs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.createBlogResponse"
                    }
                }
            }
        },
        "v1.singleBlogResponse": {
            "type": "object"
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}